+++ ydb/core/persqueue/cache_eviction.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "blob.h"
+#include <ydb/core/persqueue/pqtablet/blob/blob.h>
 #include "pq_l2_service.h"
 
 #include <ydb/core/base/appdata.h>
