+++ ydb/core/persqueue/pqtablet/blob/blob_serialization.cpp
@@ -1,7 +1,7 @@
 #include "blob.h"
 #include "blob_int.h"
-#include "header.h"
-#include "type_codecs.h"
+#include <ydb/core/persqueue/header.h>
+#include <ydb/core/persqueue/type_codecs.h>
 
 #include <util/string/builder.h>
 #include <util/string/escape.h>
