+++ ydb/core/persqueue/pqtablet/blob/ut/ya.make
@@ -1,0 +1,5 @@
+UNITTEST_FOR(ydb/core/persqueue/pqtablet/blob)
+
+SRCS(blob_ut.cpp)
+
+END()
