+++ ydb/core/persqueue/pqtablet/blob/ya.make
@@ -1,0 +1,19 @@
+LIBRARY()
+
+SRCS(
+    blob.cpp
+    blob_serialization.cpp
+)
+
+
+
+PEERDIR(
+    ydb/core/base
+    ydb/library/logger
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
