+++ ydb/core/persqueue/pqtablet/partition/partition.h
@@ -2,7 +2,7 @@
 
 #include "sourceid.h"
 
-#include <ydb/core/persqueue/blob.h>
+#include <ydb/core/persqueue/pqtablet/blob/blob.h>
 #include <ydb/core/persqueue/header.h>
 #include <ydb/core/persqueue/key.h>
 #include "partition_init.h"
