+++ ydb/core/persqueue/pqtablet/partition/partition_blob_encoder.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/core/persqueue/blob.h>
+#include <ydb/core/persqueue/pqtablet/blob/blob.h>
 #include <ydb/core/persqueue/key.h>
 #include <ydb/core/persqueue/events/internal.h>
