+++ ydb/core/persqueue/pqtablet/partition/ya.make
@@ -24,6 +24,7 @@ PEERDIR(
     ydb/core/persqueue/events
     ydb/core/persqueue/common
     ydb/core/persqueue/partition_key_range
+    ydb/core/persqueue/pqtablet/blob
     ydb/core/persqueue/pqtablet/common
     ydb/library/logger
 )
