+++ ydb/core/persqueue/ut/internals_ut.cpp
@@ -1,4 +1,4 @@
-#include "blob.h"
+#include <ydb/core/persqueue/pqtablet/blob/blob.h>
 #include "header.h"
 
 #include <library/cpp/testing/unittest/registar.h>
